% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPsychRmd.R
\name{set_jsPsych}
\alias{set_jsPsych}
\title{Set the Rmd file and directory for jsPsych}
\usage{
set_jsPsych(file_name = "task", pavlovia = FALSE)
}
\arguments{
\item{pavlovia}{If you set TRUE, set_jsPsych set Rmd for conducting the task in Pavlova.}

\item{dir_name}{directory name of jsPsych task}
}
\value{
Make directories for Research Compendium of CCP Lab and
R Markdown file for thesis at Department of Psychology, Senshu University
}
\description{
Set the Rmd file and directory for jsPsych
}
\examples{
# set_jsPsych("stroop")
}
