% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPsychRmd.R
\name{set_jsPsych}
\alias{set_jsPsych}
\title{Set the Rmd file and directory for jsPsych}
\usage{
set_jsPsych(pavlovia = FALSE, folder = FALSE)
}
\arguments{
\item{pavlovia}{If you set TRUE, set_jsPsych set Rmd for conducting the task in Pavlova (set FALSE in default)}

\item{folder}{If you set TRUE, set_jsPsych make directory that you set and put file in it(set FALSE in default)}
}
\value{
Make directories for Research Compendium of CCP Lab and
R Markdown file for thesis at Department of Psychology, Senshu University
}
\description{
Set the Rmd file and directory for jsPsych
}
\examples{
# set_jsPsych("stroop")
}
