% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsPsychRmd.R
\name{set_jsPsych}
\alias{set_jsPsych}
\title{Set Research Compendium of CCP Lab}
\usage{
set_jsPsych(file_name = "thesis", pavlovia = FALSE)
}
\arguments{
\item{file_name}{file name and directory name of RMarkdown}
}
\value{
Make directories for Research Compendium of CCP Lab and
R Markdown file for thesis at Department of Psychology, Senshu University
}
\description{
Set Research Compendium of CCP Lab
}
\examples{
# set_rc_ccp("yoshihiko_kunisato")
}
